# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "平行校园 Parallel Campus",
    "description": "知乎黑客松 2026 · 校园新锐季。公开数据 → 可编辑人格 → 分身在虚拟校园里替你先活一天。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/zhihu/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Zhihu Callback",
        "description": "回调：解析 code（兼容 `authorization_code`）→ 换 token → 建/复用 user → 302。",
        "operationId": "zhihu_callback_api_auth_zhihu_callback_get",
        "parameters": [
          {
            "name": "authorization_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization Code"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/dev-login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Dev Login",
        "description": "仅 DEV_MODE。",
        "operationId": "dev_login_api_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/judge-login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Judge Login",
        "description": "评委账号（env JUDGE_ACCOUNTS）。",
        "operationId": "judge_login_api_auth_judge_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JudgeLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/persona": {
      "get": {
        "tags": [
          "persona"
        ],
        "summary": "Get Persona",
        "operationId": "get_persona_api_persona_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "persona"
        ],
        "summary": "Put Persona",
        "operationId": "put_persona_api_persona_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonaUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampusIdentity": {
        "properties": {
          "major": {
            "type": "string",
            "maxLength": 20,
            "title": "Major",
            "default": "未定"
          },
          "grade": {
            "type": "string",
            "enum": [
              "大一",
              "大二",
              "大三",
              "大四",
              "研一",
              "研二",
              "研三"
            ],
            "title": "Grade",
            "default": "大二"
          },
          "club": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Club"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CampusIdentity"
      },
...
      "DevLoginRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Name",
            "default": "测试用户"
          }
        },
        "type": "object",
        "title": "DevLoginRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Interest": {
        "properties": {
          "topic": {
            "type": "string",
            "maxLength": 12,
            "title": "Topic"
          },
          "weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "topic",
          "weight"
        ],
        "title": "Interest"
      },
      "JudgeLoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "JudgeLoginRequest"
      },
...
      "PersonaFile": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Display Name"
          },
          "archetype": {
            "type": "string",
            "maxLength": 20,
            "title": "Archetype"
          },
          "mbti_like": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "E_I",
                "S_N",
                "T_F",
                "J_P"
              ]
            },
            "type": "object",
            "title": "Mbti Like"
          },
          "big_five": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "O",
                "C",
                "E",
                "A",
                "N"
              ]
            },
            "type": "object",
            "title": "Big Five"
          },
          "interests": {
            "items": {
              "$ref": "#/components/schemas/Interest"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 3,
            "title": "Interests"
          },
          "stances": {
            "items": {
              "$ref": "#/components/schemas/Stance"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Stances"
          },
          "speaking_style": {
            "$ref": "#/components/schemas/SpeakingStyle"
          },
          "values": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Values"
          },
          "social": {
            "$ref": "#/components/schemas/Social"
          },
          "campus_identity": {
            "$ref": "#/components/schemas/CampusIdentity"
          },
          "appearance": {
            "type": "string",
            "maxLength": 80,
            "title": "Appearance",
            "default": ""
          },
          "summary": {
            "type": "string",
            "maxLength": 400,
            "minLength": 1,
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "display_name",
          "archetype",
          "mbti_like",
          "big_five",
          "interests",
          "summary"
        ],
        "title": "PersonaFile"
      },
      "PersonaResponse": {
        "properties": {
          "file": {
            "$ref": "#/components/schemas/PersonaFile"
          },
          "thin": {
            "type": "boolean",
            "title": "Thin",
            "default": false
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "source_stats": {
            "$ref": "#/components/schemas/SourceStats"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "generated_left_today": {
            "type": "integer",
            "title": "Generated Left Today",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "PersonaResponse"
      },
      "PersonaUpdateRequest": {
        "properties": {
          "file": {
            "$ref": "#/components/schemas/PersonaFile"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "PersonaUpdateRequest"
      },
...
      "Social": {
        "properties": {
          "initiative": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Initiative",
            "default": 0.5
          },
          "group_pref": {
            "type": "string",
            "enum": [
              "独处",
              "小圈子",
              "广交"
            ],
            "title": "Group Pref",
            "default": "小圈子"
          },
          "avoid_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Avoid Topics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Social"
      },
      "SourceStats": {
        "properties": {
          "contents": {
            "type": "integer",
            "title": "Contents",
            "default": 0
          },
          "followees": {
            "type": "integer",
            "title": "Followees",
            "default": 0
          },
          "favlists": {
            "type": "integer",
            "title": "Favlists",
            "default": 0
          },
          "saved_items": {
            "type": "integer",
            "title": "Saved Items",
            "default": 0
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed"
          }
        },
        "type": "object",
        "title": "SourceStats"
      },
...
      "SpeakingStyle": {
        "properties": {
          "tone": {
            "type": "string",
            "maxLength": 30,
            "title": "Tone",
            "default": "自然"
          },
          "emoji": {
            "type": "boolean",
            "title": "Emoji",
            "default": false
          },
          "length": {
            "type": "string",
            "enum": [
              "短",
              "中",
              "长"
            ],
            "title": "Length",
            "default": "中"
          },
          "catchphrases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Catchphrases"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SpeakingStyle"
      },
      "Stance": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 40,
            "title": "Text"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "Stance"
      },
...
      "UserView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "avatar_key": {
            "type": "string",
            "title": "Avatar Key"
          },
          "auth_kind": {
            "type": "string",
            "enum": [
              "zhihu",
              "dev",
              "judge"
            ],
            "title": "Auth Kind"
          },
          "is_judge": {
            "type": "boolean",
            "title": "Is Judge",
            "default": false
          },
          "has_persona": {
            "type": "boolean",
            "title": "Has Persona",
            "default": false
          },
          "character_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "zhihu_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zhihu Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "avatar_key",
          "auth_kind"
        ],
        "title": "UserView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}